test: reduce duplication of test code
